benchmark/fixtures/template/pyproject.toml
[project]
name = "flask-api"
version = "1.0.0"
description = "Flask-based REST API service"
requires-python = ">=3.9"

dependencies = [
    "flask>=2.3.0",
    "pyjwt>=2.8.0",
    "psycopg2-binary>=2.9.0",
    "pytest>=7.4.0",
]
